+++ assets/js/theme/design-system/search.js
@@ -1,201 +1,168 @@
-/* eslint-disable no-undef */
-import { ariaHideBodyChildren } from '../utils/a11y';
-import { focusTrap } from '../utils/focus-trap';
+/* eslint-disable no-undef, no-underscore-dangle */
+window.osuny = window.osuny || {};
+window.osuny.Search = window.osuny.Search || {};
 
-const CLASSES = {
-    modalOpened: 'has-search-opened'
+window.osuny.Search = function (element) {
+    window.osuny.Popup.call(this, element);
+
+    this.buttons = document.querySelectorAll('.search-button');
+
+    this._setup();
+    this._listen();
 };
 
-class Search {
-    constructor (container) {
-        this.state = {
-            isOpened: false
-        };
-        this.buttons = document.querySelectorAll('.pagefind-ui__toggle');
-        this.toggleButton;
-        this.element = document.querySelector('.search__modal');
-        this.closeButton = this.element.querySelector('.search__close');
-        // this.a11yButton = document.querySelector('[href="#search-button"]');
+window.osuny.Search.prototype = Object.create(window.osuny.Popup.prototype);
 
-        this.searchInstance = new PagefindUI({
-            element: container,
-            showSubResults: true
-        });
+window.osuny.Search.prototype._setup = function () {
+    this.setPageFind();
+    this.setAccessibility();
+};
 
-        if (!this.element) {
-            return;
+window.osuny.Search.prototype.setPageFind = function () {
+    this.pageFindUI = new PagefindUI({
+        element: this.element,
+        showSubResults: true
+    });
+
+    // Listen to user's input
+    this.input = document.querySelector('.pagefind-ui__search-input');
+};
+
+window.osuny.Search.prototype._listen = function () {
+    var inPageWithToc = document.body.querySelector('.toc-cta');
+    window.osuny.Popup.prototype._listen.call(this);
+
+    this.buttons.forEach(function (button) {
+        button.addEventListener('click', this.toggle.bind(this, true, button));
+        if (inPageWithToc) {
+            button.classList.add('in-page-with-toc');
         }
+    }.bind(this));
 
-        this.listen();
-        this.setAccessibility();
-    }
+    this.input.addEventListener('input', this.onInputChange.bind(this));
 
-    setAccessibility () {
-        const input = document.querySelector('.pagefind-ui__search-input'),
-            // Delay before screen readers speak results message
-            speakDelay = 1500;
-        input.setAttribute('title', input.getAttribute('placeholder'));
+    this.observeMutations();
+};
 
-        // Add element to alert screen reader of the search results
-        this.accessibleMessageContainer = document.createElement('div');
-        this.element.append(this.accessibleMessageContainer);
-        this.accessibleMessageContainer.setAttribute('aria-live', 'polite');
-        this.accessibleMessageContainer.setAttribute('aria-atomic', 'true');
-        this.accessibleMessageContainer.classList.add('sr-only', 'pagefind-ui__accessible-message');
+window.osuny.Search.prototype.observeMutations = function () {
+    var previousResults = [],
+        results = [],
+        firstResult;
 
-        input.addEventListener('input', () => {
-            this.accessibleMessageContainer.innerHTML = '';
-            if (input.value !== '') {
-                this.accessibleMessage = document.createElement('p');
-                this.accessibleMessageContainer.appendChild(this.accessibleMessage);
-            }
-            clearTimeout(this.a11yTimeout);
-            this.a11yTimeout = setTimeout(this.updateAccessibility.bind(this), speakDelay);
-        });
-    }
-
-    updateAccessibility () {
-        this.updateAccessibilityMessageRole();
-        this.updateAccessibilityTags();
-    }
-
-    updateAccessibilityTags () {
-        const results = this.element.querySelectorAll('.pagefind-ui__result');
-        results.forEach(this.updateAccessibilityResult);
-    }
-
-    updateAccessibilityResult (result) {
-        let image = result.querySelector('img'),
-            title = result.querySelector('.pagefind-ui__result-title');
-
-        if (image) {
-            image.setAttribute('alt', '');
+    this.observer = new MutationObserver(function () {
+        results = this.element.querySelectorAll('.pagefind-ui__result');
+        if (results.length > 0 && previousResults.length > 0 && results.length !== previousResults.length) {
+            firstResult = results[previousResults.length];
+            setTimeout(this.focusFirstNextResult.bind(firstResult), 100);
         }
-        if (title) {
-            title.setAttribute('role', 'heading');
-            title.setAttribute('aria-level', '2');
-        }
+        previousResults = results;
+    }.bind(this));
+
+    this.observer.observe(this.element, { childList: true, subtree: true });
+};
+
+window.osuny.Search.prototype.focusFirstNextResult = function () {
+    var anchor = this.querySelector('a');
+    if (anchor) {
+        anchor.focus();
+    }
+};
+
+window.osuny.Search.prototype.onInputChange = function () {
+    // Delay before screen readers speak results message
+    var speakDelay = 1500;
+
+    // Clear aria live
+    this.accessibleMessageContainer.innerHTML = '';
+    clearTimeout(this.a11yTimeout);
+    this.a11yTimeout = setTimeout(this.updateAccessibility.bind(this), speakDelay);
+
+};
+
+window.osuny.Search.prototype.toggle = function (open, triggerElement) {
+    window.osuny.Popup.prototype.toggle.call(this, open, triggerElement);
+    if (this.state.opened) {
+        this.input.focus();
+    } else {
+        this.clear();
+    }
+};
+
+window.osuny.Search.prototype.clear = function () {
+    var button = this.element.querySelector('.pagefind-ui__button'),
+        message = this.element.querySelector('.pagefind-ui__message'),
+        results = this.element.querySelector('.pagefind-ui__results');
+
+    if (this.input) {
+        this.input.value = '';
     }
 
-    updateAccessibilityMessageRole () {
-        const message = this.element.querySelector('.pagefind-ui__message');
-        if (!message) {
-            return;
-        }
-        message.setAttribute('aria-hidden', 'true');
+    if (button) {
+        button.parentElement.removeChild(button);
+    }
+
+    if (message) {
+        message.innerText = '';
+    }
+
+    if (results) {
+        results.innerHTML = '';
+    }
+};
+
+window.osuny.Search.prototype.setAccessibility = function () {
+    this.input.setAttribute('title', this.input.getAttribute('placeholder'));
+
+    // Add element to alert screen reader of the search results
+    this.accessibleMessageContainer = document.createElement('div');
+    this.element.append(this.accessibleMessageContainer);
+    this.accessibleMessageContainer.setAttribute('aria-live', 'polite');
+    this.accessibleMessageContainer.setAttribute('aria-atomic', 'true');
+    this.accessibleMessageContainer.classList.add('sr-only', 'pagefind-ui__accessible-message');
+};
+
+window.osuny.Search.prototype.updateAccessibility = function () {
+    this.updateAccessibilityMessage();
+    this.updateAccessibilityTags();
+};
+
+window.osuny.Search.prototype.updateAccessibilityTags = function () {
+    var results = this.element.querySelectorAll('.pagefind-ui__result');
+    results.forEach(this.updateAccessibilityResult);
+};
+
+window.osuny.Search.prototype.updateAccessibilityResult = function (result) {
+    var image = result.querySelector('img'),
+        title = result.querySelector('.pagefind-ui__result-title');
+
+    if (image) {
+        image.setAttribute('alt', '');
+    }
+
+    if (title) {
+        title.setAttribute('role', 'heading');
+        title.setAttribute('aria-level', '2');
+    }
+};
+
+window.osuny.Search.prototype.updateAccessibilityMessage = function () {
+    var message = this.element.querySelector('.pagefind-ui__message');
+    if (!message) {
+        return;
+    }
+    message.setAttribute('aria-hidden', 'true');
+
+    if (this.input.value !== '') {
+        this.accessibleMessage = document.createElement('p');
         this.accessibleMessage.innerText = message.innerText;
+        this.accessibleMessageContainer.appendChild(this.accessibleMessage);
     }
-
-    listen () {
-        if (document.body.querySelector('.toc-cta')) {
-            this.buttons.forEach(button => {
-                button.classList.add('in-page-with-toc');
-            });
-        }
-        this.buttons.forEach(button => {
-            button.addEventListener('click', () => {
-                this.toggleButton = button;
-                this.toggle(true);
-                this.removedItems = this.element.querySelector('.pagefind-ui__suppressed', '.pagefind-ui__search-clear');
-                if (this.removedItems) {
-                    this.removedItems.remove();
-                }
-            });
-        });
-        this.closeButton.addEventListener('click', () => {
-            this.clearSearch();
-            this.toggle(false);
-        });
-
-        window.addEventListener('keydown', (event) => {
-            if (event.keyCode === 27 || event.key === 'Escape') {
-                if (this.state.isOpened) {
-                    this.toggle(false);
-                    this.toggleButton.focus();
-                }
-            } else if (event.key === 'Tab' && this.state.isOpened) {
-                focusTrap(event, this.element, this.state.isOpened);
-            }
-            this.buttonMore = this.element.querySelector('.pagefind-ui__results + button');
-            if (this.buttonMore && this.state.isOpened) {
-                this.buttonMore.addEventListener('click', () => {
-                    this.buttonMoreFocus();
-                });
-            }
-        });
-
-        // if (this.a11yButton) {
-        //     a11yClick(this.a11yButton, (event) => {
-        //         if (isMobile()) {
-        //             event.preventDefault();
-        //             this.toggle(true);
-        //         }
-        //     });
-        // }
-    }
-
-    clearSearch () {
-        const button = this.element.querySelector('.pagefind-ui__button'),
-            message = this.element.querySelector('.pagefind-ui__message'),
-            results = this.element.querySelector('.pagefind-ui__results');
-
-        if (this.input) {
-            this.input.value = '';
-            this.searchInstance.triggerSearch(false);
-        }
-
-        if (message) {
-            message.innerText = '';
-        }
-        if (results) {
-            results.innerHTML = '';
-        }
-        if (button) {
-            button.parentElement.removeChild(button);
-        }
-    }
-
-    buttonMoreFocus () {
-        const observer = new MutationObserver(mutations => {
-            mutations.forEach(mutation => {
-                mutation.addedNodes.forEach(addedNode => {
-                    if (addedNode instanceof HTMLAnchorElement) {
-                        addedNode.focus();
-                    }
-                });
-            });
-        });
-
-        observer.observe(this.element, { childList: true, subtree: true });
-    }
-
-    toggle(open = !this.state.isOpened, button) {
-        this.state.isOpened = open;
-        this.element.setAttribute('aria-hidden', !this.state.isOpened);
-        this.toggleButton.setAttribute('aria-expanded', this.state.isOpened);
-
-        const classAction = this.state.isOpened ? 'add' : 'remove';
-        document.documentElement.classList[classAction](CLASSES.modalOpened);
-
-        if (open) {
-            this.input = this.element.querySelector('input');
-            this.input.focus();
-        } else {
-            document.body.style.overflow = 'unset';
-            this.button.focus();
-            this.accessibleMessageContainer.innerHTML = '';
-        }
-        ariaHideBodyChildren(this.element, open);
-    }
-}
+};
 
 // Selectors
-window.addEventListener('DOMContentLoaded', () => {
-    const pageFindContainer = document.querySelector('#search');
-
-    if (typeof PagefindUI === 'undefined') return;
-
-    if (pageFindContainer) {
-        // new Search(pageFindContainer);
+window.addEventListener('DOMContentLoaded', function () {
+    var element = document.querySelector('#search');
+    if (typeof PagefindUI !== 'undefined' && element) {
+        window.osuny.search = new window.osuny.Search(element);
     }
 });
