+++ assets/js/theme/index.js
@@ -13,12 +13,11 @@ import './design-system/mainMenu';
 import './design-system/modal';
 import './design-system/navAccessibility';
 import './design-system/notes';
-import './design-system/search';
 import './design-system/toc';
 import './design-system/Popup';
 import './design-system/Lightbox';
 import './design-system/sliders';
-import './design-system/pageFind';
+import './design-system/search';
 
 import './blocks/keyFigures';
 import './blocks/timeline';
